Extract admin-only ShippingMethodType
