+++ src/Sylius/Bundle/AdminBundle/Resources/config/routing/shipping_method.yml
@@ -6,6 +6,8 @@ sylius_admin_shipping_method:
         except: ['show']
         redirect: update
         grid: sylius_admin_shipping_method
+        form:
+            type: Sylius\Bundle\AdminBundle\Form\Type\ShippingMethodType
         permission: true
     type: sylius.resource
